python/notification-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
confluent-kafka
